Merged PR 8460115: runtime save/restore support
ms-tpm-20-ref doesn't have a built-in way to save/restore global library state, so we have to inject a custom bit of code into the library to dump the contents of `Global.h`.

This code is a direct port + simplification of the eponymous file from the Windows OS repo.

With those hooks in place, the Rust wrapper can extract that state into an opaque `Vec<u8>`, and package it up alongside Rust platform state, using `serde` as a structured de/serialization engine.

To avoid leaking `serde` into the consuming lib, the save/restore state is exchanged as an opaque `Vec<u8>`, which _happens_ to be de/serialized using `postcard` - a compact, lightweight `serde`-compatible data format.

Lastly, the API has been re-jigged to allow calling `restore` at _any_ point in the TPM lib's lifetime (not just after library instantiation).

* * *

Also included in the PR is some misc cleanup (since this lib doesn't get as much housekeeping done on it as the rest of HvLite)

Related work items: #37609697